refactoring configuration and auth (#344)
## 📝 Description

**What does this PR do and why is this change necessary?**

Refactoring code related to managing the .linode-cli configuration file
Additional refactoring and tests will be done in other PRs to break up the work some
